requirements.txt
fastapi
uvicorn
anthropic
youtube-transcript-api
python-dotenv
httpx